+++ packages/image_picker/android/src/main/java/io/flutter/plugins/imagepicker/ImagePickerDelegate.java
@@ -7,6 +7,7 @@ package io.flutter.plugins.imagepicker;
 import android.Manifest;
 import android.app.Activity;
 import android.content.Intent;
+import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
 import android.content.pm.ResolveInfo;
 import android.media.MediaScannerConnection;
@@ -15,12 +16,13 @@ import android.provider.MediaStore;
 import androidx.annotation.VisibleForTesting;
 import androidx.core.app.ActivityCompat;
 import androidx.core.content.FileProvider;
-import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.PluginRegistry;
 import java.io.File;
 import java.io.IOException;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.UUID;
 
 /**
@@ -74,6 +76,14 @@ public class ImagePickerDelegate
 
   @VisibleForTesting final String fileProviderName;
 
+  static final String FLUTTER_IMAGE_PICKER_IMAGE_PATH_KEY = "flutter_image_picker_image_path";
+  static final String SHARED_PREFERENCE_ERROR_CODE_KEY = "flutter_image_picker_error_code";
+  static final String SHARED_PREFERENCE_ERROR_MESSAGE_KEY = "flutter_image_picker_error_message";
+  static final String SHARED_PREFERENCE_MAX_WIDTH_KEY = "flutter_image_picker_max_width";
+  static final String SHARED_PREFERENCE_MAX_HEIGHT_KEY = "flutter_image_picker_max_height";
+  static final String SHARED_PREFERENCE_TYPE_KEY = "flutter_image_picker_type";
+  static final String SHARED_PREFERENCE_PENDING_IMAGE_URI_PATH_KEY = "flutter_image_picker_pending_image_uri";
+
   private final Activity activity;
   private final File externalFilesDirectory;
   private final ImageResizer imageResizer;
@@ -102,9 +112,7 @@ public class ImagePickerDelegate
     void onPathReady(String path);
   }
 
-  private Uri pendingCameraMediaUri;
   private MethodChannel.Result pendingResult;
-  private MethodCall methodCall;
 
   public ImagePickerDelegate(
       final Activity activity, File externalFilesDirectory, ImageResizer imageResizer) {
@@ -112,7 +120,6 @@ public class ImagePickerDelegate
         activity,
         externalFilesDirectory,
         imageResizer,
-        null,
         null,
         new PermissionManager() {
           @Override
@@ -165,7 +172,6 @@ public class ImagePickerDelegate
       File externalFilesDirectory,
       ImageResizer imageResizer,
       MethodChannel.Result result,
-      MethodCall methodCall,
       PermissionManager permissionManager,
       IntentResolver intentResolver,
       FileUriResolver fileUriResolver,
@@ -175,18 +181,28 @@ public class ImagePickerDelegate
     this.imageResizer = imageResizer;
     this.fileProviderName = activity.getPackageName() + ".flutter.image_provider";
     this.pendingResult = result;
-    this.methodCall = methodCall;
     this.permissionManager = permissionManager;
     this.intentResolver = intentResolver;
     this.fileUriResolver = fileUriResolver;
     this.fileUtils = fileUtils;
   }
 
-  public void chooseVideoFromGallery(MethodCall methodCall, MethodChannel.Result result) {
-    if (!setPendingMethodCallAndResult(methodCall, result)) {
+  void retrieveLostImage(MethodChannel.Result result) {
+    Map resultMap = resultMapFromPendingResult();
+    if (resultMap.isEmpty()) {
+      result.success(null);
+    } else {
+      result.success(resultMap);
+    }
+    cleanPendingResultPref();
+  }
+
+  public void chooseVideoFromGallery(MethodChannel.Result result) {
+    if (!setPendingResult(result)) {
       finishWithAlreadyActiveError(result);
       return;
     }
+    setType("video");
 
     if (!permissionManager.isPermissionGranted(Manifest.permission.READ_EXTERNAL_STORAGE)) {
       permissionManager.askForPermission(
@@ -204,11 +220,12 @@ public class ImagePickerDelegate
     activity.startActivityForResult(pickVideoIntent, REQUEST_CODE_CHOOSE_VIDEO_FROM_GALLERY);
   }
 
-  public void takeVideoWithCamera(MethodCall methodCall, MethodChannel.Result result) {
-    if (!setPendingMethodCallAndResult(methodCall, result)) {
+  public void takeVideoWithCamera(MethodChannel.Result result) {
+    if (!setPendingResult(result)) {
       finishWithAlreadyActiveError(result);
       return;
     }
+    setType("video");
 
     launchTakeVideoWithCameraIntent();
   }
@@ -223,7 +240,7 @@ public class ImagePickerDelegate
     }
 
     File videoFile = createTemporaryWritableVideoFile();
-    pendingCameraMediaUri = Uri.parse("file:" + videoFile.getAbsolutePath());
+    setPendingCameraMediaUriPath("file:" + videoFile.getAbsolutePath());
 
     Uri videoUri = fileUriResolver.resolveFileProviderUriForFile(fileProviderName, videoFile);
     intent.putExtra(MediaStore.EXTRA_OUTPUT, videoUri);
@@ -232,11 +249,12 @@ public class ImagePickerDelegate
     activity.startActivityForResult(intent, REQUEST_CODE_TAKE_VIDEO_WITH_CAMERA);
   }
 
-  public void chooseImageFromGallery(MethodCall methodCall, MethodChannel.Result result) {
-    if (!setPendingMethodCallAndResult(methodCall, result)) {
+  public void chooseImageFromGallery(MethodChannel.Result result) {
+    if (!setPendingResult(result)) {
       finishWithAlreadyActiveError(result);
       return;
     }
+    setType("image");
 
     if (!permissionManager.isPermissionGranted(Manifest.permission.READ_EXTERNAL_STORAGE)) {
       permissionManager.askForPermission(
@@ -254,11 +272,12 @@ public class ImagePickerDelegate
     activity.startActivityForResult(pickImageIntent, REQUEST_CODE_CHOOSE_IMAGE_FROM_GALLERY);
   }
 
-  public void takeImageWithCamera(MethodCall methodCall, MethodChannel.Result result) {
-    if (!setPendingMethodCallAndResult(methodCall, result)) {
+  public void takeImageWithCamera(MethodChannel.Result result) {
+    if (!setPendingResult(result)) {
       finishWithAlreadyActiveError(result);
       return;
     }
+    setType("image");
 
     launchTakeImageWithCameraIntent();
   }
@@ -273,7 +292,7 @@ public class ImagePickerDelegate
     }
 
     File imageFile = createTemporaryWritableImageFile();
-    pendingCameraMediaUri = Uri.parse("file:" + imageFile.getAbsolutePath());
+    setPendingCameraMediaUriPath("file:" + imageFile.getAbsolutePath());
 
     Uri imageUri = fileUriResolver.resolveFileProviderUriForFile(fileProviderName, imageFile);
     intent.putExtra(MediaStore.EXTRA_OUTPUT, imageUri);
@@ -391,7 +410,7 @@ public class ImagePickerDelegate
   private void handleCaptureImageResult(int resultCode) {
     if (resultCode == Activity.RESULT_OK) {
       fileUriResolver.getFullImagePath(
-          pendingCameraMediaUri,
+          Uri.parse(getPendingCameraMediaUriPath()),
           new OnPathReadyListener() {
             @Override
             public void onPathReady(String path) {
@@ -408,7 +427,7 @@ public class ImagePickerDelegate
   private void handleCaptureVideoResult(int resultCode) {
     if (resultCode == Activity.RESULT_OK) {
       fileUriResolver.getFullImagePath(
-          pendingCameraMediaUri,
+              Uri.parse(getPendingCameraMediaUriPath()),
           new OnPathReadyListener() {
             @Override
             public void onPathReady(String path) {
@@ -423,42 +442,47 @@ public class ImagePickerDelegate
   }
 
   private void handleImageResult(String path, boolean shouldDeleteOriginalIfScaled) {
-    if (pendingResult != null) {
-      Double maxWidth = methodCall.argument("maxWidth");
-      Double maxHeight = methodCall.argument("maxHeight");
-
-      String finalImagePath = imageResizer.resizeImageIfNeeded(path, maxWidth, maxHeight);
+    if (pendingResult == null) {
+      saveResult(path, null, null);
+      return;
+    }
+      String finalImagePath = getFinalImagePath(path);
       finishWithSuccess(finalImagePath);
 
       //delete original file if scaled
       if (!finalImagePath.equals(path) && shouldDeleteOriginalIfScaled) {
         new File(path).delete();
       }
-    } else {
-      throw new IllegalStateException("Received image from picker that was not requested");
-    }
+  }
+
+  private String getFinalImagePath(String originalPath) {
+    Double maxWidth = getMaxWidth();
+    Double maxHeight = getMaxHeight();
+    return imageResizer.resizeImageIfNeeded(originalPath, maxWidth, maxHeight);
   }
 
   private void handleVideoResult(String path) {
-    if (pendingResult != null) {
-      finishWithSuccess(path);
-    } else {
-      throw new IllegalStateException("Received video from picker that was not requested");
-    }
+    finishWithSuccess(path);
   }
 
-  private boolean setPendingMethodCallAndResult(
-      MethodCall methodCall, MethodChannel.Result result) {
+  private boolean setPendingResult(
+     MethodChannel.Result result) {
     if (pendingResult != null) {
       return false;
     }
-
-    this.methodCall = methodCall;
     pendingResult = result;
+
+    // Clean up old pending result preference if there is a new pending result.
+    cleanPendingResultPref();
+
     return true;
   }
 
   private void finishWithSuccess(String imagePath) {
+    if (pendingResult == null) {
+      saveResult(imagePath, null, null);
+      return;
+    }
     pendingResult.success(imagePath);
     clearMethodCallAndResult();
   }
@@ -468,12 +492,100 @@ public class ImagePickerDelegate
   }
 
   private void finishWithError(String errorCode, String errorMessage) {
+    if (pendingResult == null) {
+      saveResult(null, errorCode, errorMessage);
+      return;
+    }
     pendingResult.error(errorCode, errorMessage, null);
     clearMethodCallAndResult();
   }
 
   private void clearMethodCallAndResult() {
-    methodCall = null;
     pendingResult = null;
   }
+
+  void saveResult( String path, String errorCode, String errorMessage) {
+    SharedPreferences.Editor editor =  ImagePickerPlugin.getFilePref.edit();
+    if (path != null) {
+      editor.putString(FLUTTER_IMAGE_PICKER_IMAGE_PATH_KEY, path);
+    }
+    if (errorCode != null) {
+      editor.putString(SHARED_PREFERENCE_ERROR_CODE_KEY, errorCode);
+    }
+    if (errorMessage != null) {
+      editor.putString(SHARED_PREFERENCE_ERROR_MESSAGE_KEY, errorMessage);
+    }
+    editor.apply();
+  }
+
+  void cleanPendingResultPref() {
+    ImagePickerPlugin.getFilePref.edit().clear().apply();
+  }
+
+  Map<String, String> resultMapFromPendingResult() {
+    Map<String, String> resultMap = new HashMap<>();
+    Boolean hasData = false;
+    String filePath = ImagePickerPlugin.getFilePref.getString(FLUTTER_IMAGE_PICKER_IMAGE_PATH_KEY, null);
+    if (filePath != null ) {
+      String finalPath = getFinalImagePath(filePath);
+      resultMap.put("path", finalPath);
+      hasData = true;
+    }
+    String errorCode = ImagePickerPlugin.getFilePref.getString(SHARED_PREFERENCE_ERROR_CODE_KEY, null);
+    if (errorCode != null ) {
+      resultMap.put("errorCode", errorCode);
+      hasData = true;
+      String errorMessage = ImagePickerPlugin.getFilePref.getString(SHARED_PREFERENCE_ERROR_MESSAGE_KEY, null);
+      if (errorMessage != null ) {
+        resultMap.put("errorMessage", errorMessage);
+      }
+    }
+
+    if (hasData) {
+      String type = ImagePickerPlugin.getFilePref.getString(SHARED_PREFERENCE_TYPE_KEY, null);
+      if (type != null) {
+        resultMap.put("type", type);
+      }
+    }
+    return resultMap;
+  }
+
+  Double getMaxWidth() {
+    if (!ImagePickerPlugin.getFilePref.contains(SHARED_PREFERENCE_MAX_WIDTH_KEY)) {
+      return null;
+    }
+    return Double.longBitsToDouble(ImagePickerPlugin.getFilePref.getLong(SHARED_PREFERENCE_MAX_WIDTH_KEY, 0));
+  }
+
+  Double getMaxHeight() {
+    if (!ImagePickerPlugin.getFilePref.contains(SHARED_PREFERENCE_MAX_HEIGHT_KEY)) {
+      return null;
+    }
+    return Double.longBitsToDouble(ImagePickerPlugin.getFilePref.getLong(SHARED_PREFERENCE_MAX_HEIGHT_KEY, 0));
+  }
+
+  void saveMaxDemension(Double maxWidth, Double maxHeight) {
+    SharedPreferences.Editor editor =  ImagePickerPlugin.getFilePref.edit();
+    if (maxWidth != null) {
+      editor.putLong(SHARED_PREFERENCE_MAX_WIDTH_KEY, Double.doubleToRawLongBits(maxWidth));
+    }
+    if (maxHeight != null) {
+      editor.putLong(SHARED_PREFERENCE_MAX_WIDTH_KEY, Double.doubleToRawLongBits(maxHeight));
+    }
+    editor.apply();
+  }
+
+  void setType(String type) {
+    ImagePickerPlugin.getFilePref.edit().putString(SHARED_PREFERENCE_TYPE_KEY, type).apply();
+  }
+
+  void setPendingCameraMediaUriPath(String path) {
+    ImagePickerPlugin.getFilePref.edit().putString(SHARED_PREFERENCE_PENDING_IMAGE_URI_PATH_KEY, path).apply();
+  }
+
+  String getPendingCameraMediaUriPath() {
+    return ImagePickerPlugin.getFilePref.getString(SHARED_PREFERENCE_PENDING_IMAGE_URI_PATH_KEY, "");
+  }
+
+
 }